pmdmini improvement (2608 wav samples for rythm channels) and bugfix (PZI file 64bit issue / DWORD type definition)

+++ libs/pmdmini/pmdmini/fmgen/types.h
@@ -35,7 +35,7 @@ typedef signed int int32;
 	typedef unsigned char	BYTE;
 	typedef unsigned short	WORD;
 //	typedef unsigned long	DWORD;
-	typedef unsigned long		DWORD;
+	typedef unsigned int		DWORD;
 
 
 #endif
